+++ src/common/ChatterSet.cpp
@@ -56,4 +56,9 @@ std::vector<QString> ChatterSet::filterByPrefix(const QString &prefix) const
     return result;
 }
 
+std::vector<std::pair<QString, QString>> ChatterSet::all() const
+{
+    return {this->items.begin(), this->items.end()};
+}
+
 }  // namespace chatterino