+++ src/util/Helpers.hpp
@@ -153,4 +153,6 @@ std::vector<T> splitListIntoBatches(const T &list, int batchSize = 100)
     return batches;
 }
 
+bool compareEmoteStrings(const QString &a, const QString &b);
+
 }  // namespace chatterino